D8AGE-581: Move and refactor the trait.
